9.6 webpack stuffs

+++ src/controllers/videoController.js
@@ -1,4 +1,5 @@
 import Video from "../models/Video"
+import User from "../models/User"
 
 export const home = async(req, res) => {
     try{
@@ -11,7 +12,7 @@ export const home = async(req, res) => {
 
 export const watch = async (req, res) => {
     const { id } = req.params;
-    const video = await Video.findById(id);
+    const video = await Video.findById(id).populate("owner");
     if(video){
         return res.render("watch", { pageTitle: video.title, video });
     }else{
@@ -21,19 +22,27 @@ export const watch = async (req, res) => {
 
 export const getEdit = async (req, res) => {
     const { id } = req.params;
+    const { user: {_id} } = req.session;
     const video = await Video.findById(id);
     if(!video){
         return res.status(404).render( { pageTitle: "Video Not Found." });
+    }
+    if( String(vdeo.owner) !==  String(_id)) {
+        return res.status(403).redirect("/");
     }
     return res.render("edit", { pageTitle: `Edit ${video.title}`, video });
 };
 
 export const postEdit = async (req,res) => {
     const { id } = req.params;
+    const { user: {_id} } = req.session;
     const { title, description, hashtags } = req.body;
     const video = await Video.exists({ _id: id });
     if(!video){
         return res.status(404).render( { pageTitle: "Video Not Found." });
+    }
+    if( String(vdeo.owner) !==  String(_id)) {
+        return res.status(403).redirect("/");
     }
     await Video.findByIdAndUpdate(id, {
         title,
@@ -48,17 +57,26 @@ export const getUpload = (req, res) => {
 };
 
 export const postUpload = async (req, res) => {
+    const { 
+        user: { _id },
+    } = req.session;
+    const { path: fileUrl } = req.file;
     const { title, description, hashtags } = req.body;
     try{
-        await Video.create({
+        const newVideo = await Video.create({
             title,
             description,
+            fileUrl,
+            owner: _id,
             hashtags: Video.formatHashtags(hashtags),
             meta: {
                 views: 0,
                 rating: 0,
             },
         });
+        const user = await User.findById(_id);
+        user.videos.push(newVideo._id);
+        user.save();
         return res.redirect("/");
     }catch(error){
         console.log(error);
@@ -71,6 +89,14 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
     const { id } = req.params;
+    const { user: {_id} } = req.session;
+    const video = await Video.findById(id);
+    if(!video){
+        return res.status(404).render( { pageTitle: "Video Not Found." });
+    }
+    if( String(vdeo.owner) !==  String(_id)) {
+        return res.status(403).redirect("/");
+    }
     await Video.findByIdAndDelete(id);
     return res.redirect("/");
 }
